+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Les types de dossiers sont triés par pack dans la configuration (association) d'un nouveau type de dossier #1802
 - Le script `installation/force-delete-module.php` est remplacé par la commande `app:module:force-delete-module`. Les associations aux connecteurs sont maintenant bien supprimées même s'il n'existe pas de document pour le type de dossier #1284
 - Seul l'utilisateur peut gérer ses notifications, l'administrateur n'a plus les droits #1886
-- Le script de changement d'état d'un ensemble de document (3.1) est disponible en commande Symfony `SetNewStatusDocumentBatchCommand` #1893
+- Le script de changement d'état d'un ensemble de document (3.1) est disponible en commande `app:module:set-new-status-document-batch` #1893
 
 ## Corrections